apply scalecodec patch, move to settings, update extrinsics and related subtensor calls
