+++ bittensor/core/async_subtensor.py
@@ -5219,8 +5219,8 @@ class AsyncSubtensor(SubtensorMixin):
         target_address: Optional[str] = None,
         period: Optional[int] = None,
         raise_error: bool = False,
-        wait_for_inclusion: bool = False,
-        wait_for_finalization: bool = False,
+        wait_for_inclusion: bool = True,
+        wait_for_finalization: bool = True,
     ) -> ExtrinsicResponse:
         """
         Creates a new crowdloan campaign on-chain.
