+++ bittensor/core/subtensor.py
@@ -4001,8 +4001,8 @@ class Subtensor(SubtensorMixin):
         target_address: Optional[str] = None,
         period: Optional[int] = None,
         raise_error: bool = False,
-        wait_for_inclusion: bool = False,
-        wait_for_finalization: bool = False,
+        wait_for_inclusion: bool = True,
+        wait_for_finalization: bool = True,
     ) -> ExtrinsicResponse:
         """
         Creates a new crowdloan campaign on-chain.
